boot/optee-os: Add out-of-source OPTEE-OS device tree support
Similarly to Uboot, this patch adds the ability to copy in and build
out-of-source device tree sources during an OPTEE-OS build.
To build the external device tree source file, the OP-TEE OS configuration
must refer to it with the CFG_EMBED_DTB_SOURCE_FILE option and no platform
flavor need to be specified.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>