package/gcnano-binaries: bump version to 6.4.13
Update to the latest version of the gcnano-binaries blobs, which are
compatible with the last v6.1-stm32mp-r1 kernel from the
STMicroelectronics BSP.

The location of the driver has changed it is now in the
gcnano-driver-stm32mp directory instead of an archive.

The VERSION and PREFIX temporary variable in pkgconfig file came back
therefore the sed replacement also came back.

The EULA license file has changed due to copyright year update, so its
hash has changed.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>