debugging: fix trace-cmd build
trace-cmd build can fail because of parallelized build, not finding
libtracevent.a during ctracecmd.so build
This failure happens because of multiple specific conditions:
- python is enabled in the defconfig, so ctracecmd.so is built
- ctracecmd.so recipe does not depend explicitely on libtraceevent.a
  (despite libtraceevent.a being somewhere higher in the recipe chain)
- build spawns parallel jobs

A brute solution could be to disable parallel build on trace-cmd package
(at buildroot level). Another solution is to add the missing dependency in
the sources root makefile.
The fix is relevant for this version of buildroot only: newer versions have
an updated management for trace-cmd dependencies, because libtraceevent is
built as a standalone buildroot package, and so is built before trace-cmd
package build is even started

Fixes #2

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>