kernel: enable CONFIG_BPF_JIT_ALWAYS_ON
When trying to dump the jited code for a loaded bpf program, bpftool
prints "no instructions returned"

This message is due to the fact that programs are not jited by default
(visible when using bpftool prog). It can be enabled through
/proc/sys/net/core/bpf_jit_enable, but there's no reason not to have it
systematically enabled.

Force JIT in any case with CONFIG_BPF_JIT_ALWAYS_ON.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>