patches: patch libbpf to include upstream fix
We currently get some warnings that may be converted to errors when
building the small bpf program involved in the eBPF lab. Those warnings
appear with clang >= 21.0.0. The proper way to get rid of those fixes is
to use libbpf 1.6.2, but our current buildroot version exposes libbpf
1.5.0, so bring the upstream patch.

This patch can be dropped once we update buildroot, as the package has
been bumped to 1.6.2 in 2025.11 with commit e1ef4be70b60
("package/libbpf: bump to version 1.6.2").

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>